% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov.lmr.R
\name{anova.lmr}
\alias{anova.lmr}
\title{Robust analysis of variance type tests.}
\usage{
\method{anova}{lmr}(object, ...)
}
\arguments{
\item{object, ...}{Robust linear models fit by \code{robustbase::lmrob}.}
}
\description{
Robust analysis of variance type tests.
}
\details{
The first object must be the model with the least terms.The results of
  the robust F-test and robust chi-squared test are both returned. The second
  of these should, generally, be preferred. The function is a simple
  wrapper to \link{\code{lmrobLinTest}}.
}
