% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LMR.R
\name{consoleLMR}
\alias{consoleLMR}
\title{consoleLMR}
\usage{
consoleLMR(waveMin, waveMax, waveStep, coreD, L, dL, Layers, angleMax)
}
\arguments{
\item{waveMin}{Wavelength start [um].}

\item{waveMax}{Wavelength end [um].}

\item{waveStep}{Wavelength step [um].}

\item{coreD}{Core diameter [um].}

\item{L}{Length of the modified region [um].}

\item{dL}{Thickness of the coating layers [nm].}

\item{Layers}{Layers.}

\item{angleMax}{Maximum skewness angle [deg].}
}
\value{
Data frame with the wavelength and real part of the transmittance.
}
\description{
Theoretical modelling of lossy mode resonance (LMR)
}
\details{
consoleLMR
}
\examples{
\dontrun{
waveMin <- 0.5
waveMax <- 1.5
waveStep <- 0.01
temperature <- 300
coreD <- 200
L <- 25000
angleMax <- 90
thickLayers <- c( 1165 )

Core <- function( lambda, temperature ){

  a1 <- 0.6961663
  a2 <- 0.4079426
  a3 <- 0.8974794
  b1 <- 0.0684043
  b2 <- 0.1162414
  b3 <- 9.896161
  core <- sqrt(1.0+(a1*lambda^2/(lambda^2-b1^2))+
  (a2*lambda^2/(lambda^2-b2^2))+(a3*lambda^2/(lambda^2-b3^2)))

  thermal_drift <- 1+1.28e-5 * temperature
  thermal_drift <- 1
  res <- core * thermal_drift

  return( res )

}

L1 <- function( lambda, temperature ){

  lambda <- lambda*10^(-6)
  epsinf <- 1.0
  Ak <- 101.0
  Bk <- 1.2
  Ek <- 6.2
  hplanck <- 6.62607004*10^(-34)
  vlight <- 3*10^8
  ladel <- 1.60217733*10^(-19)
  E <- ( hplanck * vlight / lambda ) / ladel
  eps <- epsinf + Ak / ( Ek^2 - E^2 - 1i * Bk * E )
  res <- sqrt(eps)

  thermal_drift <- 1.0 - 1.49e-4 * temperature
  thermal_drift <- 1.0
  res <- res * thermal_drift

  return( res )

}

Layers <- data.frame( lambda = seq( waveMin, waveMax, by = waveStep ) )
Layers$Cladding <- 1.43
Layers$Sensing <- 1.321
Layers$Core <- Core( Layers$lambda, temperature )
Layers$Layers1 <- L1( Layers$lambda, temperature )

Result <- consoleLMR( waveMin, waveMax, waveStep, coreD, L, thickLayers, Layers, angleMax )
}
}
