% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmr.R
\name{frb}
\alias{frb}
\title{Fast robust bootstrap}
\usage{
frb(lmrob.object, nboot = 1000, return.coef = FALSE)
}
\arguments{
\item{lmrob.object, nboot, return.coef}{As for \code{FRB::frb}.}
}
\description{
Fast robust bootstrap
}
\details{
This is a simple wrapper for \code{FRB::frb}.
}
