% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.lmr.R
\name{plot.lmr}
\alias{plot.lmr}
\alias{ggplot.lmr}
\title{ggplot for rlm results}
\usage{
\method{plot}{lmr}(x = NULL, hist.scale = 10, plot. = TRUE, caption = NULL, ...)
}
\arguments{
\item{hist.scale}{A scaleing factor for bin widths in the histogram. The default bin width
will be \code{range(resid(data))/hist.scale)}. Defaults to \code{hist.scale=10},
which is smaler than the \code{ggplot} default of 30.}

\item{plot.}{Logical indicating whether to plot. If FALSE, the function
returns a list of ggplot objects that can be passed to \code{gridExtra::grid.arrange}.}

\item{caption}{A caption to appear at the bottom left of the plot}

\item{...}{Additional arguments passed to \code{ggplot}. Currently unused.}

\item{data}{an object of class \code{rlm}.}
}
\description{
ggplot for rlm results
}
