% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmr.R
\name{py}
\alias{py}
\title{Peno-Yohai regression estimates}
\usage{
py(
  formula,
  data,
  resid_keep_method = "proportion",
  psc_keep = 0.5,
  resid_keep_prop = 0.5,
  ...
)
}
\arguments{
\item{formula}{A formula describing a linear model.}

\item{data}{A data frame.}

\item{resid_keep_method, psc_keep, resid_keep_prop, ...}{All passed through to
\code{pyinit::pyinit}. Basically, this just provides defaults that
\code{pyinit::pyinit} insists the user decides upon.}
}
\description{
Peno-Yohai regression estimates
}
\details{
This is a simple wrapper to \code{pyinit::pyinit} that imposes
  some defaults and provides a formula/data interface.
}
