% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefInd.R
\docType{data}
\name{RefInd}
\alias{RefInd}
\title{List of refractive indices}
\format{
A data frame with 67 rows and 4 columns:
\describe{
  \item{Name}{}
  \item{lambda}{}
  \item{Index}{}
  \item{Info}{}
}
}
\source{
<https://en.wikipedia.org/wiki/List_of_refractive_indices>
}
\usage{
RefInd
}
\description{
List of refractive indices
}
\keyword{datasets}
