% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmr.R
\name{rfpe}
\alias{rfpe}
\title{Robust final predcition error}
\usage{
rfpe(object, scale)
}
\arguments{
\item{object}{An object of class 'lmr'.}

\item{scale}{The scale estimate. This must be provided: RFPE is not comparable
across models with different scale estimates. The scale estimate should be
the one from the model with the full set of predictors.}
}
\description{
Robust final predcition error
}
\details{
This is a modified copy-paste of \code{RobStatTM::lmrobdetMM.RFPE}.
}
