% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApp.R
\name{runShinyLMR}
\alias{runShinyLMR}
\title{runShinyLMR}
\usage{
runShinyLMR(launch.browser = TRUE)
}
\arguments{
\item{launch.browser}{By default Shiny application is opened in the default browser.}
}
\description{
Function to run Shiny application
}
\details{
runShinyLMR
}
\examples{
\dontrun{
runShinyLMR()
}

}
