% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmr.R
\name{rfpe}
\alias{rfpe}
\title{Robust final predcition error}
\usage{
rfpe(object, scale = NULL)
}
\arguments{
\item{object}{An object of class 'lmr'.}

\item{scale}{An optional scale parameter. Defaults to \code{scale = NULL}.}
}
\description{
Robust final predcition error
}
\details{
This is a modified copy-paste of \code{RobStatTM::lmrobdetMM.RFPE}.
}
